financeai/backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
anthropic
openai
google-genai
pydantic
httpx
python-dateutil
